tokenInfo取得の関数化 #21 / prefectureIdの設定
