+++ server/script/TransferFromEventLogs.ts
@@ -1,22 +1,16 @@
-import { ethers } from "ethers";
 import { ALCHEMY_API_KEY, NETWORK } from "@/config/project";
 import {
   getProvider,
   getLocalNounsTokenContract,
-  getLocalNounsProviderContract,
 } from "@/utils/const";
 import { addresses } from "@/utils/addresses";
 import { writeTokenDataToFirestore } from "./tokenOnFirestore";
+import { getTokenInfo } from "./tokenOnContract";
 import { TOKEN } from "@/firestore/const";
 
 const provider = getProvider(NETWORK, ALCHEMY_API_KEY);
 const tokenContract = getLocalNounsTokenContract(
   addresses.localNounsToken[NETWORK],
-  provider,
-);
-
-const providerContract = getLocalNounsProviderContract(
-  addresses.localProvider[NETWORK],
   provider,
 );
 
@@ -35,40 +29,12 @@ const main = async () => {
       // eventからto, tokenIdを取得
       const { to, tokenId } = event.args;
 
-      // if(tokenId <= 1019){
-      //   continue;
-      // }
+      if(tokenId >= 5){
+        continue;
+      }
 
-      // traits情報を取得
-      const jsonString = await providerContract.generateTraits(tokenId);
-      const traits = JSON.parse("[" + jsonString + "]");
-
-      // SVGイメージを取得(base64でエンコードされているのでデコードする)
-      // HTML表示時のサイズ調整のためwidth,heiht属性も削除する
-      const [svgPart] = await providerContract.generateSVGPart(tokenId);
-      const svg = Buffer.from(svgPart, "base64")
-        .toString("utf8")
-        .replace(/ width="320" height="320"/, "");
-
-      const { prefecture, head, accessory } = convertTrais(traits);
-      const salePrice = await tokenContract.getPriceOf(tokenId);
-      const ethPrice = ethers.formatEther(salePrice);
-      const holder = await tokenContract.ownerOf(tokenId);
-      const isOnTrade = await tokenContract.trades(tokenId);
-      const tradeToPrefecture = await tokenContract.getTradePrefectureFor(tokenId);
-
-      const tokenInfo: TOKEN = {
-        tokenId: tokenId,
-        prefecture: prefecture,
-        head: head,
-        accessory: accessory,
-        holder: holder.toLowerCase(), // firestoreでfilterするために小文字変換
-        svg: svg,
-        salePrice: Number(ethPrice),
-        isOnTrade: isOnTrade,
-        tradeToPrefecture: tradeToPrefecture.length > 0 ? tradeToPrefecture : [0],
-        createdDate: new Date(),
-      };
+      // token情報を取得ß
+      const tokenInfo: TOKEN = await getTokenInfo(tokenId);
 
       // firestoreに書き込み
       await writeTokenDataToFirestore(tokenInfo);
